+++ src/Raytracing/Materials/Metal.h
@@ -7,10 +7,10 @@ namespace Materials
 	class Metal : public Material
 	{
 	public:
-		__device__ Metal(const glm::dvec3& albedo, double fuzz) : albedo(albedo), fuzz(fuzz < 1 ? fuzz : 1) { }
+		__device__ Metal(const glm::dvec3& albedo, float fuzz) : albedo(albedo), fuzz(fuzz < 1 ? fuzz : 1) { }
 
-		__device__ bool scatter(const Ray& rayIn, const hitData& data, glm::dvec3& attenuation, Ray& rayScattered, curandState* localRandState) const {
-			glm::dvec3 reflected = Utils::Vector::reflect(rayIn.direction(), data.normal);
+		__device__ bool scatter(const Ray& rayIn, const hitData& data, glm::vec3& attenuation, Ray& rayScattered, curandState* localRandState) const {
+			glm::vec3 reflected = Utils::Vector::reflect(rayIn.direction(), data.normal);
 			reflected = glm::normalize(reflected) + (fuzz * Utils::Vector::randomInUnitSphereVector(localRandState));
 
 			rayScattered = Ray(data.p, reflected);
@@ -20,7 +20,7 @@ namespace Materials
 		}
 
 	private:
-		glm::dvec3 albedo;
-		double fuzz;
+		glm::vec3 albedo;
+		float fuzz;
 	};
 }
